Fold OpenGL blur into a BlurAlgorithm, reuse PreDrawBlurController
